# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/sessions": {
      "post": {
        "description": "Cria uma sessão e a inicia. O próximo passo é buscar o QR code em `GET /v1/sessions/{id}/qr` e escaneá-lo com o WhatsApp do celular. Assim que o número conectar, `phoneNumber` é preenchido automaticamente.",
        "operationId": "sessions_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credencial ausente, inválida ou revogada.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "A chave não possui o escopo necessário para esta operação.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "A sessão não existe ou pertence a outra aplicação.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos. O campo `errors` detalha por campo.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido. Veja o header `Retry-After`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno. Informe o `requestId` ao acionar o suporte.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Criar uma sessão",
        "tags": [
          "Sessões"
        ]
      },
      "get": {
        "operationId": "sessions_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "STOPPED",
                "STARTING",
                "SCAN_QR_CODE",
                "PASSKEY_REQUIRED",
                "PASSKEY_CONFIRMATION_REQUIRED",
                "WORKING",
                "FAILED",
                "UNKNOWN"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca por apelido ou número.",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credencial ausente, inválida ou revogada.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "A chave não possui o escopo necessário para esta operação.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "A sessão não existe ou pertence a outra aplicação.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos. O campo `errors` detalha por campo.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido. Veja o header `Retry-After`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno. Informe o `requestId` ao acionar o suporte.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Listar suas sessões",
        "tags": [
          "Sessões"
        ]
      }
    },
...
    "/v1/sessions/{id}/events": {
      "get": {
        "description": "Fluxo Server-Sent Events com as mudanças de estado e as mensagens da sessão. Útil para saber o instante em que o QR foi lido, sem ficar consultando.",
        "operationId": "sessionEvents_events",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Credencial ausente, inválida ou revogada.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "A chave não possui o escopo necessário para esta operação.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos. O campo `errors` detalha por campo.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido. Veja o header `Retry-After`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno. Informe o `requestId` ao acionar o suporte.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Acompanhar uma sessão em tempo real",
        "tags": [
          "Sessões"
        ]
      }
    },
...
  },
  "info": {
    "title": "WhatsApp Gateway W51",
...
    "version": "1.0",
    "contact": {},
    "license": {
      "name": "Proprietário",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://{host}",
      "description": "Produção",
      "variables": {
        "host": {
          "default": "gateway.seu-dominio.com",
          "description": "Domínio da sua instalação."
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave da sua aplicação, emitida no painel administrativo."
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Identificador estável do tipo de erro, no padrão recurso/motivo.",
            "example": "https://gateway.w51/errors/session-not-found"
          },
          "title": {
            "type": "string",
            "description": "Resumo curto do erro.",
            "example": "Sessão não encontrada"
          },
          "status": {
            "type": "number",
            "description": "Código HTTP.",
            "example": 404
          },
          "detail": {
            "type": "string",
            "description": "Explicação em português, voltada a quem vai corrigir o problema.",
            "example": "Nenhuma sessão com o id informado pertence à sua aplicação."
          },
          "instance": {
            "type": "string",
            "description": "Caminho que originou o erro.",
            "example": "/v1/sessions/abc123"
          },
          "requestId": {
            "type": "string",
            "description": "Id da requisição — informe-o ao pedir suporte, é por ele que achamos o log.",
            "example": "01JCQ8Z5X9K2M4N6P8R0T2V4W6"
          },
          "errors": {
            "type": "object",
            "description": "Erros de validação por campo, quando houver.",
            "example": {
              "chatId": [
                "Número precisa incluir o código do país."
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "requestId"
        ]
      },
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "description": "Apelido para você identificar este número. Único dentro da sua aplicação. O nome técnico da sessão é gerado por nós.",
            "example": "Comercial"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Dados livres seus, devolvidos nas consultas e nos webhooks.",
            "example": {
              "setor": "vendas",
              "responsavel": "time-1"
            }
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1a2b3c4d5e6f7g8h9"
          },
          "label": {
            "type": "string",
            "nullable": true,
            "example": "Comercial"
          },
          "status": {
            "type": "string",
            "enum": [
              "STOPPED",
              "STARTING",
              "SCAN_QR_CODE",
              "PASSKEY_REQUIRED",
              "PASSKEY_CONFIRMATION_REQUIRED",
              "WORKING",
              "FAILED",
              "UNKNOWN"
            ],
            "description": "SCAN_QR_CODE = aguardando leitura do QR · WORKING = conectado e pronto para enviar."
          },
          "statusLabel": {
            "type": "string",
            "example": "Conectado",
            "description": "Status em português."
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "description": "Preenchido quando o número conecta.",
            "example": "5511999999999"
          },
          "pushName": {
            "type": "string",
            "nullable": true,
            "example": "Comercial da Empresa"
          },
          "engine": {
            "type": "string",
            "example": "NOWEB"
          },
          "qrRequestCount": {
            "type": "number",
            "description": "Quantas vezes o QR code desta sessão foi solicitado.",
            "example": 2
          },
          "lastQrRequestedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "connectedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Quando o número foi conectado."
          },
          "disconnectedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "description": "Seus dados livres."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "statusLabel",
          "engine",
          "qrRequestCount",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}